% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmt.meta.data.df.R
\name{frmt.meta.data.df}
\alias{frmt.meta.data.df}
\title{Format NEON AQU Site Hydrologic Metadata}
\arguments{
\item{input.list}{Required: list created by shiny that contains the site ID, start date, 
and end date selected by the user [list]}

\item{product.list}{Required: list stored as a reference table that contains the metadata
needed to build the data frame [data frame]}
}
\value{
Returns a data frame.
}
\description{
This function formats reach-scale hydrologic and morphological metadata for a
NEON AQU site into a dataframe to be rendered in the discharge visualization shiny app UI.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Zachary L. Nickerson \email{nickerson@battelleecology.org} \cr
}
