% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.GET.R
\name{pheno.GET}
\alias{pheno.GET}
\title{Get URL to render Phenocam image for a NEON site at a given timestamp}
\arguments{
\item{dpID}{Required: Phenocam DPID formatted as DP#.##### [string]}

\item{siteID}{Required: NEON site ID [string]}

\item{domainID}{Required: NEON domain ID [string]}

\item{dateTime}{Required: Date-time formatted as YYYY-MM-DDTHH:MM:SSZ [string]}
}
\value{
This function returns a url to render a Phenocam image. If no image available, the
function returns NULL.
}
\description{
This function queries the Phenocam API to retrieve an image for the nearest
timestamp basedo on user input and prints the image as an output.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
James Ross \email{ross.james94@gmail.com} \cr
}
