% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.modal.R
\name{pheno.modal}
\alias{pheno.modal}
\title{Phenocam modal}
\arguments{
\item{phenoURL}{Required: URL generated by pheno.GET [string]}

\item{usrDateTime}{Required: Formatted date and time for bad modal [string]}

\item{isGoodRequest}{Required: Boolean generated by pheno.GET [boolean]}

\item{siteID}{Required: SiteID from app [string]}
}
\value{
This function returns nothing
}
\description{
Displays appropriate modal for phenoGET functionality.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
James Ross \email{ross.james94@gmail.com} \cr
}
