% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.modal.R
\name{phenoModal}
\alias{phenoModal}
\title{Phenocam modal}
\value{
This function returns nothing
}
\description{
Displays appropriate modal for phenoGET functionality.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
James Ross \email{ross.james94@gmail.com} \cr
}
