% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.modal.R
\name{phenoModalBad}
\alias{phenoModalBad}
\title{Phenocam modal bad request}
\usage{
phenoModalBad(usrDateTime, siteID)
}
\value{
This function returns a modal object based on phenoGET return.
}
\description{
Displays appropriate bad request modal for phenoGET functionality.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
James Ross \email{ross.james94@gmail.com} \cr
}
