% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.modal.R
\name{phenoModalGood}
\alias{phenoModalGood}
\title{Phenocam modal good request}
\usage{
phenoModalGood(phenoURL)
}
\value{
This function returns a modal object based on phenoGET return.
}
\description{
Displays appropriate good request modal for phenoGET functionality.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
James Ross \email{ross.james94@gmail.com} \cr
}
