% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cont.Q.R
\name{plot.cont.Q}
\alias{plot.cont.Q}
\title{Plot NEON Continuous Discharge (DP4.00130.001) Data}
\arguments{
\item{site.id}{Required: NEON AQU site ID selected by the shiny app user [string]}

\item{start.date}{Required: Search interval start date (YYYY-MM-DD) selected by the shiny 
app user [string]}

\item{end.date}{Required: Search interval end date (YYYY-MM-DD) selected by the shiny app 
user [string]}

\item{input.list}{Required: List containing the data used in plotting [list]}

\item{plot.final.QF}{Required: Indicator of plotting the finalDischargeQF field from the 
DP4.00130.001 data product [boolean]}

\item{plot.sci.rvw.QF}{Required: Indicator of plotting the finalDischargeQFSciRvw field 
from the DP4.00130.001 data product [boolean]}
}
\value{
Returns a plotly plot object
}
\description{
This function will generate a plotly plot of NEON Continuous Discharge
(DP4.00130.001) data to be rendered in the discharge visualization shiny app
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Zachary L. Nickerson \email{nickerson@battelleecology.org} \cr
}
