% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RC.R
\name{plot.RC}
\alias{plot.RC}
\title{Plot NEON Stage-Discharge Rating Curves (DP4.00133.001) Data}
\arguments{
\item{site.id}{Required: NEON AQU site ID selected by the shiny app user [string]}

\item{start.date}{Required: Search interval start date (YYYY-MM-DD) selected by the shiny
app user [string]}

\item{end.date}{Required: Search interval end date (YYYY-MM-DD) selected by the shiny app
user [string]}

\item{input.list}{Required: List containing the curve IDs used in plotting [list]}

\item{plot.imp.unit}{Required: Idicator of plotting data in metric or imperial units
[boolean]}

\item{mode.dark}{Required: Indicator of plotting data in light or dark mode [boolean]}
}
\value{
Returns a plotly plot object
}
\description{
This function will generate a plotly plot of NEON Stage-Discharge Rating
Curves (DP4.00130.001) data to be rendered in the discharge visualization shiny app
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Zachary L. Nickerson \email{nickerson@battelleecology.org} \cr
James M. Ross \email{ross.james94@gmail.com} \cr
}
