% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RC.plot.R
\name{RC.plot}
\alias{RC.plot}
\title{Plot NEON Stage-Discharge Rating Curves (DP4.00133.001) Data}
\arguments{
\item{site.id}{Required: NEON AQU site ID selected by the shiny app user [string]}

\item{start.date}{Required: Search interval start date (YYYY-MM-DD) selected by the shiny
app user [string]}

\item{end.date}{Required: Search interval end date (YYYY-MM-DD) selected by the shiny app
user [string]}

\item{input.list}{Required: List containing the data used in plotting; this list should be
the outputs from neonStageQPlot::get.cont.Q.NEON.API [list]}

\item{plot.imp.unit}{Defaults to FALSE: Idicator of plotting data in metric or imperial
units [boolean]}

\item{mode.dark}{Defaults to FALSE: Indicator of plotting data in light or dark mode. NOTE:
This argument is only used in the openFlow shiny app. [boolean]}
}
\value{
Returns a plotly plot object
}
\description{
This function will generate a plotly plot of NEON Stage-Discharge Rating
Curves (DP4.00130.001) data to be rendered in the openFlow shiny app. The function can
also be used outside the app, but should only be used with the outputs from
neonStageQPlot::get.cont.Q.NEON.API.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Zachary L. Nickerson \email{nickerson@battelleecology.org} \cr
James M. Ross \email{ross.james94@gmail.com} \cr
}
