% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.RC.cont.Q.plot.R
\name{run.RC.cont.Q.plot}
\alias{run.RC.cont.Q.plot}
\title{NEON continuous discharge and stage-discharge rating curve data visualization app}
\value{
This function launches a shiny app to interactively plot published NEON data
}
\description{
An interactive app which  plots continuous discharge, Stage discharge and
rating curves using language R Data is downloaded from the NEON data portal using
neonUltilities and ploting is done at 20mins time interval
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Divine Aseaku \email{divineaseaku@gmail.com} \cr
Zachary L. Nickerson \email{nickerson@battelleecology.org} \cr
James Ross \email{ross.james94@gmail.com} \cr
}
