% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.cont.Q.NEON.API.R
\name{get.cont.Q.NEON.API}
\alias{get.cont.Q.NEON.API}
\title{Download and Wrangle DP4.00130.001 and DP4.00133.001 Data from NEON Portal API}
\arguments{
\item{site.id}{Required: NEON AQU site ID selected by the shiny app user [string]}

\item{start.date}{Required: Search interval start date (YYYY-MM-DD) selected by the shiny
app user [string]}

\item{end.date}{Required: Search interval end date (YYYY-MM-DD) selected by the shiny app
user [string]}

\item{api.token}{Optional: NEON API token provided by the user}
}
\value{
Returns a list of:
1) 'continuousDischarge_sum' is a data frame that contains continuous stage and discharge at
20 min intervals, discrete stage and discharge, and quality flags (finalDischargeQF and
dischargeFinalQFSciRvw from DP4.00130.001), and
2) 'curveIDs' is a vector of all the unique rating curve IDs that are used to build the
timeseries data.
}
\description{
This function will download data from the NEON Portal API using neonUtilities
based on user input and wrangle the data to create a summary table smoothed to 20 min
temporal resolution.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Zachary L. Nickerson \email{nickerson@battelleecology.org} \cr
James M. Ross \email{ross.james94@gmail.com} \cr
}
