% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.GET.R
\name{pheno.GET}
\alias{pheno.GET}
\title{Phenocam GET request}
\arguments{
\item{siteID}{Required: SiteID from app [string]}

\item{domainID}{Required: domainID from app [string]}

\item{dateTime}{Required: dateTime generated from plotly click event [string]}
}
\value{
This function returns a url to a phenocam image. If no image available returns NULL.
}
\description{
Uses given siteID, domainID, and dateTime to retrieve url to a phenocam image.
Closest to given parameters.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
James Ross \email{ross.james94@gmail.com} \cr
}
